Created abstracted edit campaign form and replaced create form with it.
