+++ ui/components/cards/CampaignCards.tsx
@@ -45,7 +45,7 @@ const CampaignCardWrapper: FunctionComponent<CampaignCardWrapperProps> = ({
           "sm:flex-row": !forceColumn,
         })}
       >
-        <CampaignImage imageUrl={campaign.profileImageUrl} />
+        <CampaignImage url={campaign.profileImageUrl} />
         <div
           className={cn(
             "flex flex-col items-stretch flex-1",
