+++ ui/components/index.ts
@@ -16,6 +16,7 @@ export * from "./Carousel"
 export * from "./CenteredColumn"
 export * from "./ContributeForm"
 export * from "./ContributionGraph"
+export * from "./EditCampaignForm"
 export * from "./FAQQuestion"
 export * from "./Footer"
 export * from "./Header"
